pyproject.toml
[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "app"]

[tool.pyright]
include = ["app"]
[project]
name = "fastapi-to-do-list"
version = "0.1.0"
description = ""
authors = [
    {name = "sap2204",email = "zufs@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic (==1.15.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "async-timeout (==5.0.1)",
    "asyncpg (==0.30.0)",
    "autoflake (==2.3.1)",
    "bcrypt (==4.3.0)",
    "black (==25.1.0)",
    "certifi (==2025.1.31)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "dnspython (==2.7.0)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.2.0)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.12)",
    "fastapi-cli (==0.0.7)",
    "flake8 (==7.2.0)",
    "greenlet (==3.1.1)",
    "gunicorn (==23.0.0)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "iniconfig (==2.1.0)",
    "isort (==6.0.1)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.6)",
    "mako (==1.3.9)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mccabe (==0.7.0)",
    "mdurl (==0.1.2)",
    "mypy-extensions (==1.0.0)",
    "nodeenv (==1.9.1)",
    "orjson (==3.10.16)",
    "packaging (==24.2)",
    "passlib (==1.7.4)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.7)",
    "pluggy (==1.5.0)",
    "pyasn1 (==0.4.8)",
    "pycodestyle (==2.13.0)",
    "pydantic (==2.11.1)",
    "pydantic-extra-types (==2.10.3)",
    "pydantic-settings (==2.8.1)",
    "pydantic-core (==2.33.0)",
    "pyflakes (==3.3.2)",
    "pygments (==2.19.1)",
    "pyright (==1.1.399)",
    "pytest (==8.3.5)",
    "pytest-asyncio (==0.26.0)",
    "python-dotenv (==1.1.0)",
    "python-jose (==3.4.0)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "rich (==14.0.0)",
    "rich-toolkit (==0.14.1)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.40)",
    "starlette (==0.46.1)",
    "tomli (==2.2.1)",
    "typer (==0.15.2)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.0)",
    "ujson (==5.10.0)",
    "uvicorn (==0.34.0)",
    "watchfiles (==1.0.4)",
    "websockets (==15.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
